+++ ledger/api-server-damlonx/reference-v2/BUILD.bazel
@@ -45,6 +45,7 @@ da_scala_binary(
         "//ledger/participant-state",
         "//ledger/participant-state/kvutils",
         "//ledger/sandbox",
+        "//libs-scala/resources",
         "@maven//:com_github_scopt_scopt_2_12",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
         "@maven//:com_typesafe_akka_akka_stream_2_12",
