+++ ledger/sandbox/BUILD.bazel
@@ -57,6 +57,7 @@ da_scala_library(
         "//ledger/participant-state",
         "//ledger/participant-state-index",
         "//libs-scala/direct-execution-context",
+        "//libs-scala/resources",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
         "@maven//:com_github_scopt_scopt_2_12",
@@ -131,6 +132,7 @@ da_scala_library(
         "//ledger/participant-state",
         "//ledger/participant-state-index",
         "//libs-scala/direct-execution-context",
+        "//libs-scala/resources",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
         "@maven//:com_github_scopt_scopt_2_12",
@@ -256,6 +258,7 @@ da_scala_library(
         "//ledger/test-common",
         "//libs-scala/direct-execution-context",
         "//libs-scala/grpc-utils",
+        "//libs-scala/resources",
         "//libs-scala/timer-utils",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
@@ -310,6 +313,7 @@ da_scala_test_suite(
         "//ledger/test-common",
         "//libs-scala/direct-execution-context",
         "//libs-scala/grpc-utils",
+        "//libs-scala/resources",
         "//libs-scala/timer-utils",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
