+++ libs-scala/resources/BUILD.bazel
@@ -1,0 +1,28 @@
+# Copyright (c) 2020 The DAML Authors. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load("//bazel_tools:scala.bzl", "da_scala_library", "da_scala_test_suite")
+
+da_scala_library(
+    name = "resources",
+    srcs = glob(["src/main/scala/**/*.scala"]),
+    tags = ["maven_coordinates=com.digitalasset:resources:__VERSION__"],
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [
+        "@maven//:com_typesafe_akka_akka_actor_2_12",
+        "@maven//:com_typesafe_akka_akka_stream_2_12",
+        "@maven//:org_scala_lang_modules_scala_java8_compat_2_12",
+    ],
+)
+
+da_scala_test_suite(
+    name = "resources-tests",
+    srcs = glob(["src/test/suite/**/*.scala"]),
+    deps = [
+        ":resources",
+        "@maven//:com_typesafe_akka_akka_actor_2_12",
+        "@maven//:com_typesafe_akka_akka_stream_2_12",
+    ],
+)
