+++ libs-scala/resources/src/main/scala/com/digitalasset/resources/ExecutorServiceResourceOwner.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.resources
+package com.digitalasset.resources
 
 import java.util.concurrent.{ExecutorService, TimeUnit}
